feat: :sparkles: when exiting and returning to the page restore the flow
